feat(locate): Extend two jumps search until search domain is found
This is supposed to improve the locating algorithm. The two jumps used before
may not cover the target date (see, for example,
tests/test_locate.py::TestGapCase3::test_S4), which is required to extend a
search domain afterward.

Adds a "look-around" jump-based search step to find a target segment or outline
a search domain for the next step.